# docs/api/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
...
paths:
...
  /api/v1/auth/reset-password:
    post:
      tags:
      - Authentication
      summary: Reset user password
      description: Allows the user to set a new password using a valid recovery code.
      operationId: resetPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
        required: true
      responses:
        "400":
          description: Invalid or expired recovery code
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "200":
          description: Password successfully reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
              example:
                status: SUCCESS
                message: Password restored
  /api/v1/auth/register:
    post:
      tags:
      - Authentication
      summary: Register new user
      description: Creates a new user account in the system.
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "400":
          description: Invalid registration data
        "409":
          description: User already exists in the database
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "200":
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
              example:
                status: SUCCESS
                message: User registered successfully
  /api/v1/auth/refresh:
    post:
      tags:
      - Authentication
      summary: Refresh JWT token
      description: Generates new JWT tokens using the refresh token stored in cookies.
      operationId: refreshToken
      parameters:
      - name: RefreshToken
        in: cookie
        description: Refresh token (from cookie named RefreshToken)
        required: false
        schema:
          type: string
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "401":
          description: Invalid or expired refresh token
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "200":
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
              example:
                status: SUCCESS
                message: Tokens refreshed
  /api/v1/auth/logout:
    post:
      tags:
      - Authentication
      summary: User logout
      description: Clears authentication cookies and invalidates session tokens.
      operationId: logOut
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "200":
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
              example:
                status: SUCCESS
                message: logout successfully
  /api/v1/auth/login:
    post:
      tags:
      - Authentication
      summary: User login
      description: Authenticates the user and returns JWT tokens (access and refresh).
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "401":
          description: Invalid credentials
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
              example:
                status: SUCCESS
                message: Auth successful. Tokens are created in cookie.
  /api/v1/auth/forgot-password:
    post:
      tags:
      - Authentication
      summary: Request password recovery
      description: Sends a recovery email to the user containing a reset link or verification
        code.
      operationId: forgotPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
        required: true
      responses:
        "400":
          description: Invalid email address
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "404":
          description: There is no user registered with this email
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "200":
          description: Recovery email sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
              example:
                status: SUCCESS
                message: Recovery email sent successfully
...
  /api/v1/auth/me:
    get:
      tags:
      - Authentication
      summary: Get the active user
      description: Obtain details of the currently authenticated user.
      operationId: getActiveUser
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "401":
          description: User not authenticated
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "200":
          description: User successfully authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRole"
              example:
                username: johndoe
                roles:
                - USER
...
  /api/v1/auth:
    delete:
      tags:
      - Authentication
      summary: Delete the user's account
      description: Delete the account of the active user from the system.
      operationId: deleteAccount
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "404":
          description: User not found
          content:
            application/json: {}
        "401":
          description: Unauthorized access
          content:
            application/json: {}
        "500":
          description: Internal server error
          content:
            application/json: {}
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "200":
          description: Account successfully removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
components:
  schemas:
...
    AuthResponse:
      type: object
      properties:
        status:
          type: string
          enum:
          - SUCCESS
          - FAILURE
        message:
          type: string
        error:
          type: string
...
    ResetPasswordRequest:
      type: object
      properties:
        code:
          type: string
          pattern: "\\d{4}"
        newPassword:
          type: string
      required:
      - code
      - newPassword
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
        username:
          type: string
        password:
          type: string
      required:
      - email
      - password
      - username
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
      - password
      - username
    ForgotPasswordRequest:
      type: object
      properties:
        email:
          type: string
      required:
      - email
...
    UserRole:
      type: object
      properties:
        username:
          type: string
        roles:
          type: array
          items:
            type: string
...